# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gazin",
        "description": "Documentação da API Gazin. Esta API gerencia níveis e desenvolvedores.",
        "contact": {
            "email": "lfp_costa@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/desenvolvedores": {
            "get": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Lista os desenvolvedores com paginação e busca.",
                "description": "Retorna uma lista de desenvolvedores com suporte a paginação e busca por termo no campo 'nome'.",
                "operationId": "4c4478bb4ffac6cb911aa4a3ef26bb03",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca para filtrar os desenvolvedores pelo campo 'nome'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de desenvolvedores retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID do desenvolvedor.",
                                                        "type": "integer"
                                                    },
                                                    "nome": {
                                                        "description": "Nome do desenvolvedor.",
                                                        "type": "string"
                                                    },
                                                    "nivel_id": {
                                                        "description": "ID do nível associado.",
                                                        "type": "integer"
                                                    },
                                                    "sexo": {
                                                        "description": "Sexo do desenvolvedor.",
                                                        "type": "string"
                                                    },
                                                    "data_nascimento": {
                                                        "description": "Data de nascimento do desenvolvedor.",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "hobby": {
                                                        "description": "Hobby do desenvolvedor.",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "description": "Total de registros encontrados.",
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "description": "Página atual.",
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "description": "Número de itens por página.",
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "description": "Última página disponível.",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nenhum registro encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhum registro encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Cria um novo desenvolvedor.",
                "description": "Cria um novo registro de desenvolvedor com base nos dados fornecidos.",
                "operationId": "a3bdbd42c682a13deeed7e8ff58fa929",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nivel_id": {
                                        "description": "ID do nível associado.",
                                        "type": "integer"
                                    },
                                    "nome": {
                                        "description": "Nome do desenvolvedor.",
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "description": "Sexo do desenvolvedor.",
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "description": "Data de nascimento do desenvolvedor.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "hobby": {
                                        "description": "Hobby do desenvolvedor.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Desenvolvedor criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID do desenvolvedor criado.",
                                                    "type": "integer"
                                                },
                                                "nivel_id": {
                                                    "description": "ID do nível associado.",
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "description": "Nome do desenvolvedor.",
                                                    "type": "string"
                                                },
                                                "sexo": {
                                                    "description": "Sexo do desenvolvedor.",
                                                    "type": "string"
                                                },
                                                "data_nascimento": {
                                                    "description": "Data de nascimento do desenvolvedor.",
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "hobby": {
                                                    "description": "Hobby do desenvolvedor.",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "É necessário haver um nível cadastrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não é possível cadastrar desenvolvedor. Não existem níveis cadastrados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/desenvolvedores/{id}": {
            "put": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Atualiza um desenvolvedor existente.",
                "description": "Atualiza os dados de um desenvolvedor com base no ID fornecido.",
                "operationId": "579346e77f0604bf0209d32f045f5fc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do desenvolvedor a ser atualizado.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nivel_id": {
                                        "description": "ID do nível associado.",
                                        "type": "integer"
                                    },
                                    "nome": {
                                        "description": "Nome do desenvolvedor.",
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "description": "Sexo do desenvolvedor.",
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "description": "Data de nascimento do desenvolvedor.",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "hobby": {
                                        "description": "Hobby do desenvolvedor.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Desenvolvedor atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID do desenvolvedor atualizado.",
                                                    "type": "integer"
                                                },
                                                "nivel_id": {
                                                    "description": "ID do nível associado.",
                                                    "type": "integer"
                                                },
                                                "nome": {
                                                    "description": "Nome do desenvolvedor.",
                                                    "type": "string"
                                                },
                                                "sexo": {
                                                    "description": "Sexo do desenvolvedor.",
                                                    "type": "string"
                                                },
                                                "data_nascimento": {
                                                    "description": "Data de nascimento do desenvolvedor.",
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "hobby": {
                                                    "description": "Hobby do desenvolvedor.",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "É necessário haver um nível cadastrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não é possível cadastrar desenvolvedor. Não existem níveis cadastrados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Desenvolvedores"
                ],
                "summary": "Remove um desenvolvedor.",
                "description": "Remove um desenvolvedor com base no ID fornecido.",
                "operationId": "040653c8ff1d3f2c5137853a52abbc90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do desenvolvedor a ser removido.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Desenvolvedor removido com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro removido com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao tentar remover o desenvolvedor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao tentar remover o desenvolvedor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}